added coverage for issue 736 test case 1
